+++ crates/usvg/src/text/layout.rs
@@ -1,46 +1,64 @@
-use crate::text::old::{
-    chunk_span_at, convert_decoration, is_word_separator_characters, process_anchor,
-    resolve_baseline, script_supports_letter_spacing, shape_text, span_contains, ByteIndex,
-    DatabaseExt, DecorationSpan, FontsCache, Glyph, GlyphClusters, PathNormal, ResolvedFont,
-};
 use crate::tree::{BBox, IsValidLength};
 use crate::{
-    ApproxZeroUlps, Fill, FillRule, Font, FontStretch, FontStyle, Group, LengthAdjust, Node,
-    PaintOrder, Path, ShapeRendering, Stroke, Text, TextChunk, TextFlow, TextPath, TextRendering,
-    TextSpan, Visibility, WritingMode,
+    AlignmentBaseline, ApproxZeroUlps, BaselineShift, DominantBaseline, Fill, FillRule, Font,
+    FontStretch, FontStyle, LengthAdjust, PaintOrder, Path, ShapeRendering, Stroke, Text,
+    TextAnchor, TextChunk, TextDecorationStyle, TextFlow, TextPath, TextSpan, Visibility,
+    WritingMode,
 };
-use fontdb::ID;
+use fontdb::{Database, ID};
 use kurbo::{ParamCurve, ParamCurveArclen, ParamCurveDeriv};
+use rustybuzz::ttf_parser;
 use rustybuzz::ttf_parser::GlyphId;
 use std::collections::HashMap;
-use std::mem::transmute;
+use std::num::NonZeroU16;
 use std::sync::Arc;
 use strict_num::NonZeroPositiveF32;
 use svgtypes::FontFamily;
 use tiny_skia_path::{NonZeroRect, Transform};
 use unicode_script::UnicodeScript;
 
-#[derive(Clone, Debug)]
+/// A glyph that has already been positioned correctly.
+///
+/// Note that the transform already takes the font size into consideration, so applying the
+/// transform to the outline of the glyphs is all that is necessary to display it correctly.
+#[derive(Clone, Copy, Debug)]
 pub struct PositionedGlyph {
+    /// The transform of the glyph. This transform should be applied to the _glyph outlines_, meaning
+    /// that paint servers referenced by the glyph's span should not be affected by it.
     pub transform: Transform,
+    /// The ID of the glyph.
     pub glyph_id: GlyphId,
+    /// The ID of the font the glyph should be taken from.
     pub font: ID,
-    byte_idx: ByteIndex,
 }
 
+/// A span contains a number of layouted glyphs that share the same fill, stroke, paint order and
+/// visibility.
 #[derive(Clone, Debug)]
-pub struct PositionedSpan {
+pub struct Span {
+    /// The fill of the span.
     pub fill: Option<Fill>,
+    /// The stroke of the span.
     pub stroke: Option<Stroke>,
+    /// The paint order of the span.
     pub paint_order: PaintOrder,
+    /// The font size of the span.
     pub font_size: NonZeroPositiveF32,
+    /// The visibility of the span.
     pub visibility: Visibility,
+    /// The glyphs that make up the span.
     pub positioned_glyphs: Vec<PositionedGlyph>,
 }
 
+/// A text fragment is a component of a text element.
+///
+/// It can either be a span (which contains a number of layouted glyphs that share
+/// some properties) or a path (which is a layouted decoration span).
 #[derive(Clone, Debug)]
-pub enum PositionedTextFragment {
-    Span(PositionedSpan),
+pub enum TextFragment {
+    /// A text span.
+    Span(Span),
+    /// A path.
     Path(Path),
 }
 
@@ -85,7 +103,7 @@ pub(crate) fn convert(text: &mut Text, fontdb: &fontdb::Database) -> Option<()> 
 fn layout_text(
     text_node: &Text,
     fontdb: &fontdb::Database,
-) -> Option<(Vec<PositionedTextFragment>, NonZeroRect)> {
+) -> Option<(Vec<TextFragment>, NonZeroRect)> {
     let mut fonts_cache: FontsCache = HashMap::new();
 
     for chunk in &text_node.chunks {
@@ -169,7 +187,7 @@ fn layout_text(
                     convert_decoration(offset, span, font, decoration, &decoration_spans, span_ts)
                 {
                     bbox = bbox.expand(path.data.bounds());
-                    text_fragments.push(PositionedTextFragment::Path(path));
+                    text_fragments.push(TextFragment::Path(path));
                 }
             }
 
@@ -183,7 +201,7 @@ fn layout_text(
                     convert_decoration(offset, span, font, decoration, &decoration_spans, span_ts)
                 {
                     bbox = bbox.expand(path.data.bounds());
-                    text_fragments.push(PositionedTextFragment::Path(path));
+                    text_fragments.push(TextFragment::Path(path));
                 }
             }
 
@@ -197,10 +215,10 @@ fn layout_text(
                 fill.rule = FillRule::NonZero;
             }
 
-            if let Some((mut span_fragments, span_bbox)) = convert_span(span, &clusters, span_ts) {
+            if let Some((span_fragments, span_bbox)) = convert_span(span, &clusters, span_ts) {
                 bbox = bbox.expand(span_bbox);
 
-                let mut positioned_glyphs = span_fragments
+                let positioned_glyphs = span_fragments
                     .into_iter()
                     .flat_map(|mut gc| {
                         let cluster_ts = gc.transform();
@@ -211,7 +229,7 @@ fn layout_text(
                     })
                     .collect();
 
-                text_fragments.push(PositionedTextFragment::Span(PositionedSpan {
+                text_fragments.push(TextFragment::Span(Span {
                     fill,
                     stroke: span.stroke.clone(),
                     paint_order: span.paint_order,
@@ -231,7 +249,7 @@ fn layout_text(
                     convert_decoration(offset, span, font, decoration, &decoration_spans, span_ts)
                 {
                     bbox = bbox.expand(path.data.bounds());
-                    text_fragments.push(PositionedTextFragment::Path(path));
+                    text_fragments.push(TextFragment::Path(path));
                 }
             }
         }
@@ -324,6 +342,64 @@ fn collect_decoration_spans(span: &TextSpan, clusters: &[GlyphCluster]) -> Vec<D
     spans
 }
 
+pub(crate) fn convert_decoration(
+    dy: f32,
+    span: &TextSpan,
+    font: &ResolvedFont,
+    mut decoration: TextDecorationStyle,
+    decoration_spans: &[DecorationSpan],
+    transform: Transform,
+) -> Option<Path> {
+    debug_assert!(!decoration_spans.is_empty());
+
+    let thickness = font.underline_thickness(span.font_size.get());
+
+    let mut builder = tiny_skia_path::PathBuilder::new();
+    for dec_span in decoration_spans {
+        let rect = match NonZeroRect::from_xywh(0.0, -thickness / 2.0, dec_span.width, thickness) {
+            Some(v) => v,
+            None => {
+                log::warn!("a decoration span has a malformed bbox");
+                continue;
+            }
+        };
+
+        let ts = dec_span.transform.pre_translate(0.0, dy);
+
+        let mut path = tiny_skia_path::PathBuilder::from_rect(rect.to_rect());
+        path = match path.transform(ts) {
+            Some(v) => v,
+            None => continue,
+        };
+
+        builder.push_path(&path);
+    }
+
+    let mut path_data = builder.finish()?;
+    path_data = path_data.transform(transform)?;
+
+    Path::new(
+        String::new(),
+        span.visibility,
+        decoration.fill.take(),
+        decoration.stroke.take(),
+        PaintOrder::default(),
+        ShapeRendering::default(),
+        Arc::new(path_data),
+        Transform::default(),
+    )
+}
+
+/// A text decoration span.
+///
+/// Basically a horizontal line, that will be used for underline, overline and line-through.
+/// It doesn't have a height, since it depends on the Font metrics.
+#[derive(Clone, Copy)]
+pub(crate) struct DecorationSpan {
+    pub(crate) width: f32,
+    pub(crate) transform: Transform,
+}
+
 /// Resolves clusters positions.
 ///
 /// Mainly sets the `transform` property.
@@ -393,6 +469,50 @@ fn resolve_clusters_positions_horizontal(
     }
 
     (x, y)
+}
+
+// Baseline resolving in SVG is a mess.
+// Not only it's poorly documented, but as soon as you start mixing
+// `dominant-baseline` and `alignment-baseline` each application/browser will produce
+// different results.
+//
+// For now, resvg simply tries to match Chrome's output and not the mythical SVG spec output.
+//
+// See `alignment_baseline_shift` method comment for more details.
+pub(crate) fn resolve_baseline(
+    span: &TextSpan,
+    font: &ResolvedFont,
+    writing_mode: WritingMode,
+) -> f32 {
+    let mut shift = -resolve_baseline_shift(&span.baseline_shift, font, span.font_size.get());
+    println!("{:?}", font.units_per_em);
+
+    // TODO: support vertical layout as well
+    if writing_mode == WritingMode::LeftToRight {
+        if span.alignment_baseline == AlignmentBaseline::Auto
+            || span.alignment_baseline == AlignmentBaseline::Baseline
+        {
+            shift += font.dominant_baseline_shift(span.dominant_baseline, span.font_size.get());
+        } else {
+            shift += font.alignment_baseline_shift(span.alignment_baseline, span.font_size.get());
+        }
+    }
+
+    shift
+}
+
+fn resolve_baseline_shift(baselines: &[BaselineShift], font: &ResolvedFont, font_size: f32) -> f32 {
+    let mut shift = 0.0;
+    for baseline in baselines.iter().rev() {
+        match baseline {
+            BaselineShift::Baseline => {}
+            BaselineShift::Subscript => shift -= font.subscript_offset(font_size),
+            BaselineShift::Superscript => shift += font.superscript_offset(font_size),
+            BaselineShift::Number(n) => shift += n,
+        }
+    }
+
+    shift
 }
 
 fn resolve_clusters_positions_path(
@@ -477,6 +597,20 @@ fn resolve_clusters_positions_path(
     }
 
     (last_x, last_y)
+}
+
+pub(crate) fn process_anchor(a: TextAnchor, text_width: f32) -> f32 {
+    match a {
+        TextAnchor::Start => 0.0, // Nothing.
+        TextAnchor::Middle => -text_width / 2.0,
+        TextAnchor::End => -text_width,
+    }
+}
+
+pub(crate) struct PathNormal {
+    pub(crate) x: f32,
+    pub(crate) y: f32,
+    pub(crate) angle: f32,
 }
 
 fn collect_normals(
@@ -860,7 +994,6 @@ fn form_glyph_clusters(glyphs: &[Glyph], text: &str, font_size: f32) -> GlyphClu
             transform: ts,
             font: glyph.font.id,
             glyph_id: glyph.id,
-            byte_idx: glyph.byte_idx,
         });
 
         x += glyph.width as f32;
@@ -945,10 +1078,597 @@ fn resolve_font(font: &Font, fontdb: &fontdb::Database) -> Option<ResolvedFont> 
     fontdb.load_font(id?)
 }
 
-fn resolve_rendering_mode(text: &Text) -> ShapeRendering {
-    match text.rendering_mode {
-        TextRendering::OptimizeSpeed => ShapeRendering::CrispEdges,
-        TextRendering::OptimizeLegibility => ShapeRendering::GeometricPrecision,
-        TextRendering::GeometricPrecision => ShapeRendering::GeometricPrecision,
+pub(crate) trait DatabaseExt {
+    fn load_font(&self, id: ID) -> Option<ResolvedFont>;
+    fn has_char(&self, id: ID, c: char) -> bool;
+}
+
+impl DatabaseExt for Database {
+    #[inline(never)]
+    fn load_font(&self, id: ID) -> Option<ResolvedFont> {
+        self.with_face_data(id, |data, face_index| -> Option<ResolvedFont> {
+            let font = ttf_parser::Face::parse(data, face_index).ok()?;
+
+            let units_per_em = NonZeroU16::new(font.units_per_em())?;
+
+            let ascent = font.ascender();
+            let descent = font.descender();
+
+            let x_height = font
+                .x_height()
+                .and_then(|x| u16::try_from(x).ok())
+                .and_then(NonZeroU16::new);
+            let x_height = match x_height {
+                Some(height) => height,
+                None => {
+                    // If not set - fallback to height * 45%.
+                    // 45% is what Firefox uses.
+                    u16::try_from((f32::from(ascent - descent) * 0.45) as i32)
+                        .ok()
+                        .and_then(NonZeroU16::new)?
+                }
+            };
+
+            let line_through = font.strikeout_metrics();
+            let line_through_position = match line_through {
+                Some(metrics) => metrics.position,
+                None => x_height.get() as i16 / 2,
+            };
+
+            let (underline_position, underline_thickness) = match font.underline_metrics() {
+                Some(metrics) => {
+                    let thickness = u16::try_from(metrics.thickness)
+                        .ok()
+                        .and_then(NonZeroU16::new)
+                        // `ttf_parser` guarantees that units_per_em is >= 16
+                        .unwrap_or_else(|| NonZeroU16::new(units_per_em.get() / 12).unwrap());
+
+                    (metrics.position, thickness)
+                }
+                None => (
+                    -(units_per_em.get() as i16) / 9,
+                    NonZeroU16::new(units_per_em.get() / 12).unwrap(),
+                ),
+            };
+
+            // 0.2 and 0.4 are generic offsets used by some applications (Inkscape/librsvg).
+            let mut subscript_offset = (units_per_em.get() as f32 / 0.2).round() as i16;
+            let mut superscript_offset = (units_per_em.get() as f32 / 0.4).round() as i16;
+            if let Some(metrics) = font.subscript_metrics() {
+                subscript_offset = metrics.y_offset;
+            }
+
+            if let Some(metrics) = font.superscript_metrics() {
+                superscript_offset = metrics.y_offset;
+            }
+
+            Some(ResolvedFont {
+                id,
+                units_per_em,
+                ascent,
+                descent,
+                x_height,
+                underline_position,
+                underline_thickness,
+                line_through_position,
+                subscript_offset,
+                superscript_offset,
+            })
+        })?
+    }
+
+    #[inline(never)]
+    fn has_char(&self, id: ID, c: char) -> bool {
+        let res = self.with_face_data(id, |font_data, face_index| -> Option<bool> {
+            let font = ttf_parser::Face::parse(font_data, face_index).ok()?;
+            font.glyph_index(c)?;
+            Some(true)
+        });
+
+        res == Some(Some(true))
+    }
+}
+
+/// Text shaping with font fallback.
+pub(crate) fn shape_text(
+    text: &str,
+    font: Arc<ResolvedFont>,
+    small_caps: bool,
+    apply_kerning: bool,
+    fontdb: &fontdb::Database,
+) -> Vec<Glyph> {
+    let mut glyphs = shape_text_with_font(text, font.clone(), small_caps, apply_kerning, fontdb)
+        .unwrap_or_default();
+
+    // Remember all fonts used for shaping.
+    let mut used_fonts = vec![font.id];
+
+    // Loop until all glyphs become resolved or until no more fonts are left.
+    'outer: loop {
+        let mut missing = None;
+        for glyph in &glyphs {
+            if glyph.is_missing() {
+                missing = Some(glyph.byte_idx.char_from(text));
+                break;
+            }
+        }
+
+        if let Some(c) = missing {
+            let fallback_font = match find_font_for_char(c, &used_fonts, fontdb) {
+                Some(v) => Arc::new(v),
+                None => break 'outer,
+            };
+
+            // Shape again, using a new font.
+            let fallback_glyphs = shape_text_with_font(
+                text,
+                fallback_font.clone(),
+                small_caps,
+                apply_kerning,
+                fontdb,
+            )
+            .unwrap_or_default();
+
+            let all_matched = fallback_glyphs.iter().all(|g| !g.is_missing());
+            if all_matched {
+                // Replace all glyphs when all of them were matched.
+                glyphs = fallback_glyphs;
+                break 'outer;
+            }
+
+            // We assume, that shaping with an any font will produce the same amount of glyphs.
+            // This is incorrect, but good enough for now.
+            if glyphs.len() != fallback_glyphs.len() {
+                break 'outer;
+            }
+
+            // TODO: Replace clusters and not glyphs. This should be more accurate.
+
+            // Copy new glyphs.
+            for i in 0..glyphs.len() {
+                if glyphs[i].is_missing() && !fallback_glyphs[i].is_missing() {
+                    glyphs[i] = fallback_glyphs[i].clone();
+                }
+            }
+
+            // Remember this font.
+            used_fonts.push(fallback_font.id);
+        } else {
+            break 'outer;
+        }
+    }
+
+    // Warn about missing glyphs.
+    for glyph in &glyphs {
+        if glyph.is_missing() {
+            let c = glyph.byte_idx.char_from(text);
+            // TODO: print a full grapheme
+            log::warn!(
+                "No fonts with a {}/U+{:X} character were found.",
+                c,
+                c as u32
+            );
+        }
+    }
+
+    glyphs
+}
+
+/// Converts a text into a list of glyph IDs.
+///
+/// This function will do the BIDI reordering and text shaping.
+fn shape_text_with_font(
+    text: &str,
+    font: Arc<ResolvedFont>,
+    small_caps: bool,
+    apply_kerning: bool,
+    fontdb: &fontdb::Database,
+) -> Option<Vec<Glyph>> {
+    fontdb.with_face_data(font.id, |font_data, face_index| -> Option<Vec<Glyph>> {
+        let rb_font = rustybuzz::Face::from_slice(font_data, face_index)?;
+
+        let bidi_info = unicode_bidi::BidiInfo::new(text, Some(unicode_bidi::Level::ltr()));
+        let paragraph = &bidi_info.paragraphs[0];
+        let line = paragraph.range.clone();
+
+        let mut glyphs = Vec::new();
+
+        let (levels, runs) = bidi_info.visual_runs(paragraph, line);
+        for run in runs.iter() {
+            let sub_text = &text[run.clone()];
+            if sub_text.is_empty() {
+                continue;
+            }
+
+            let hb_direction = if levels[run.start].is_rtl() {
+                rustybuzz::Direction::RightToLeft
+            } else {
+                rustybuzz::Direction::LeftToRight
+            };
+
+            let mut buffer = rustybuzz::UnicodeBuffer::new();
+            buffer.push_str(sub_text);
+            buffer.set_direction(hb_direction);
+
+            let mut features = Vec::new();
+            if small_caps {
+                features.push(rustybuzz::Feature::new(
+                    rustybuzz::Tag::from_bytes(b"smcp"),
+                    1,
+                    ..,
+                ));
+            }
+
+            if !apply_kerning {
+                features.push(rustybuzz::Feature::new(
+                    rustybuzz::Tag::from_bytes(b"kern"),
+                    0,
+                    ..,
+                ));
+            }
+
+            let output = rustybuzz::shape(&rb_font, &features, buffer);
+
+            let positions = output.glyph_positions();
+            let infos = output.glyph_infos();
+
+            for (pos, info) in positions.iter().zip(infos) {
+                let idx = run.start + info.cluster as usize;
+                debug_assert!(text.get(idx..).is_some());
+
+                glyphs.push(Glyph {
+                    byte_idx: ByteIndex::new(idx),
+                    id: GlyphId(info.glyph_id as u16),
+                    dx: pos.x_offset,
+                    dy: pos.y_offset,
+                    width: pos.x_advance,
+                    font: font.clone(),
+                });
+            }
+        }
+
+        Some(glyphs)
+    })?
+}
+
+/// Finds a font with a specified char.
+///
+/// This is a rudimentary font fallback algorithm.
+fn find_font_for_char(
+    c: char,
+    exclude_fonts: &[fontdb::ID],
+    fontdb: &fontdb::Database,
+) -> Option<ResolvedFont> {
+    let base_font_id = exclude_fonts[0];
+
+    // Iterate over fonts and check if any of them support the specified char.
+    for face in fontdb.faces() {
+        // Ignore fonts, that were used for shaping already.
+        if exclude_fonts.contains(&face.id) {
+            continue;
+        }
+
+        // Check that the new face has the same style.
+        let base_face = fontdb.face(base_font_id)?;
+        if base_face.style != face.style
+            && base_face.weight != face.weight
+            && base_face.stretch != face.stretch
+        {
+            continue;
+        }
+
+        if !fontdb.has_char(face.id, c) {
+            continue;
+        }
+
+        let base_family = base_face
+            .families
+            .iter()
+            .find(|f| f.1 == fontdb::Language::English_UnitedStates)
+            .unwrap_or(&base_face.families[0]);
+
+        let new_family = face
+            .families
+            .iter()
+            .find(|f| f.1 == fontdb::Language::English_UnitedStates)
+            .unwrap_or(&base_face.families[0]);
+
+        log::warn!("Fallback from {} to {}.", base_family.0, new_family.0);
+        return fontdb.load_font(face.id);
+    }
+
+    None
+}
+
+/// An iterator over glyph clusters.
+///
+/// Input:  0 2 2 2 3 4 4 5 5
+/// Result: 0 1     4 5   7
+pub(crate) struct GlyphClusters<'a> {
+    data: &'a [Glyph],
+    idx: usize,
+}
+
+impl<'a> GlyphClusters<'a> {
+    pub(crate) fn new(data: &'a [Glyph]) -> Self {
+        GlyphClusters { data, idx: 0 }
+    }
+}
+
+impl<'a> Iterator for GlyphClusters<'a> {
+    type Item = (std::ops::Range<usize>, ByteIndex);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.idx == self.data.len() {
+            return None;
+        }
+
+        let start = self.idx;
+        let cluster = self.data[self.idx].byte_idx;
+        for g in &self.data[self.idx..] {
+            if g.byte_idx != cluster {
+                break;
+            }
+
+            self.idx += 1;
+        }
+
+        Some((start..self.idx, cluster))
+    }
+}
+
+/// Checks that selected script supports letter spacing.
+///
+/// [In the CSS spec](https://www.w3.org/TR/css-text-3/#cursive-tracking).
+///
+/// The list itself is from: https://github.com/harfbuzz/harfbuzz/issues/64
+pub(crate) fn script_supports_letter_spacing(script: unicode_script::Script) -> bool {
+    use unicode_script::Script;
+
+    !matches!(
+        script,
+        Script::Arabic
+            | Script::Syriac
+            | Script::Nko
+            | Script::Manichaean
+            | Script::Psalter_Pahlavi
+            | Script::Mandaic
+            | Script::Mongolian
+            | Script::Phags_Pa
+            | Script::Devanagari
+            | Script::Bengali
+            | Script::Gurmukhi
+            | Script::Modi
+            | Script::Sharada
+            | Script::Syloti_Nagri
+            | Script::Tirhuta
+            | Script::Ogham
+    )
+}
+
+/// A glyph.
+///
+/// Basically, a glyph ID and it's metrics.
+#[derive(Clone)]
+pub(crate) struct Glyph {
+    /// The glyph ID in the font.
+    pub(crate) id: GlyphId,
+
+    /// Position in bytes in the original string.
+    ///
+    /// We use it to match a glyph with a character in the text chunk and therefore with the style.
+    pub(crate) byte_idx: ByteIndex,
+
+    /// The glyph offset in font units.
+    pub(crate) dx: i32,
+
+    /// The glyph offset in font units.
+    pub(crate) dy: i32,
+
+    /// The glyph width / X-advance in font units.
+    pub(crate) width: i32,
+
+    /// Reference to the source font.
+    ///
+    /// Each glyph can have it's own source font.
+    pub(crate) font: Arc<ResolvedFont>,
+}
+
+impl Glyph {
+    fn is_missing(&self) -> bool {
+        self.id.0 == 0
+    }
+}
+
+#[derive(Clone, Copy, Debug)]
+pub(crate) struct ResolvedFont {
+    pub(crate) id: ID,
+
+    units_per_em: NonZeroU16,
+
+    // All values below are in font units.
+    ascent: i16,
+    descent: i16,
+    x_height: NonZeroU16,
+
+    underline_position: i16,
+    underline_thickness: NonZeroU16,
+
+    // line-through thickness should be the the same as underline thickness
+    // according to the TrueType spec:
+    // https://docs.microsoft.com/en-us/typography/opentype/spec/os2#ystrikeoutsize
+    line_through_position: i16,
+
+    subscript_offset: i16,
+    superscript_offset: i16,
+}
+
+pub(crate) fn chunk_span_at(chunk: &TextChunk, byte_offset: ByteIndex) -> Option<&TextSpan> {
+    chunk
+        .spans
+        .iter()
+        .find(|&span| span_contains(span, byte_offset))
+}
+
+pub(crate) fn span_contains(span: &TextSpan, byte_offset: ByteIndex) -> bool {
+    byte_offset.value() >= span.start && byte_offset.value() < span.end
+}
+
+/// Checks that the selected character is a word separator.
+///
+/// According to: https://www.w3.org/TR/css-text-3/#word-separator
+pub(crate) fn is_word_separator_characters(c: char) -> bool {
+    matches!(
+        c as u32,
+        0x0020 | 0x00A0 | 0x1361 | 0x010100 | 0x010101 | 0x01039F | 0x01091F
+    )
+}
+
+impl ResolvedFont {
+    #[inline]
+    pub(crate) fn scale(&self, font_size: f32) -> f32 {
+        font_size / self.units_per_em.get() as f32
+    }
+
+    #[inline]
+    pub(crate) fn ascent(&self, font_size: f32) -> f32 {
+        self.ascent as f32 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub(crate) fn descent(&self, font_size: f32) -> f32 {
+        self.descent as f32 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub(crate) fn height(&self, font_size: f32) -> f32 {
+        self.ascent(font_size) - self.descent(font_size)
+    }
+
+    #[inline]
+    pub(crate) fn x_height(&self, font_size: f32) -> f32 {
+        self.x_height.get() as f32 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub(crate) fn underline_position(&self, font_size: f32) -> f32 {
+        self.underline_position as f32 * self.scale(font_size)
+    }
+
+    #[inline]
+    fn underline_thickness(&self, font_size: f32) -> f32 {
+        self.underline_thickness.get() as f32 * self.scale(font_size)
+    }
+
+    #[inline]
+    pub(crate) fn line_through_position(&self, font_size: f32) -> f32 {
+        self.line_through_position as f32 * self.scale(font_size)
+    }
+
+    #[inline]
+    fn subscript_offset(&self, font_size: f32) -> f32 {
+        self.subscript_offset as f32 * self.scale(font_size)
+    }
+
+    #[inline]
+    fn superscript_offset(&self, font_size: f32) -> f32 {
+        self.superscript_offset as f32 * self.scale(font_size)
+    }
+
+    fn dominant_baseline_shift(&self, baseline: DominantBaseline, font_size: f32) -> f32 {
+        let alignment = match baseline {
+            DominantBaseline::Auto => AlignmentBaseline::Auto,
+            DominantBaseline::UseScript => AlignmentBaseline::Auto, // unsupported
+            DominantBaseline::NoChange => AlignmentBaseline::Auto,  // already resolved
+            DominantBaseline::ResetSize => AlignmentBaseline::Auto, // unsupported
+            DominantBaseline::Ideographic => AlignmentBaseline::Ideographic,
+            DominantBaseline::Alphabetic => AlignmentBaseline::Alphabetic,
+            DominantBaseline::Hanging => AlignmentBaseline::Hanging,
+            DominantBaseline::Mathematical => AlignmentBaseline::Mathematical,
+            DominantBaseline::Central => AlignmentBaseline::Central,
+            DominantBaseline::Middle => AlignmentBaseline::Middle,
+            DominantBaseline::TextAfterEdge => AlignmentBaseline::TextAfterEdge,
+            DominantBaseline::TextBeforeEdge => AlignmentBaseline::TextBeforeEdge,
+        };
+
+        self.alignment_baseline_shift(alignment, font_size)
+    }
+
+    // The `alignment-baseline` property is a mess.
+    //
+    // The SVG 1.1 spec (https://www.w3.org/TR/SVG11/text.html#BaselineAlignmentProperties)
+    // goes on and on about what this property suppose to do, but doesn't actually explain
+    // how it should be implemented. It's just a very verbose overview.
+    //
+    // As of Nov 2022, only Chrome and Safari support `alignment-baseline`. Firefox isn't.
+    // Same goes for basically every SVG library in existence.
+    // Meaning we have no idea how exactly it should be implemented.
+    //
+    // And even Chrome and Safari cannot agree on how to handle `baseline`, `after-edge`,
+    // `text-after-edge` and `ideographic` variants. Producing vastly different output.
+    //
+    // As per spec, a proper implementation should get baseline values from the font itself,
+    // using `BASE` and `bsln` TrueType tables. If those tables are not present,
+    // we have to synthesize them (https://drafts.csswg.org/css-inline/#baseline-synthesis-fonts).
+    // And in the worst case scenario simply fallback to hardcoded values.
+    //
+    // Also, most fonts do not provide `BASE` and `bsln` tables to begin with.
+    //
+    // Again, as of Nov 2022, Chrome does only the latter:
+    // https://github.com/chromium/chromium/blob/main/third_party/blink/renderer/platform/fonts/font_metrics.cc#L153
+    //
+    // Since baseline TrueType tables parsing and baseline synthesis are pretty hard,
+    // we do what Chrome does - use hardcoded values. And it seems like Safari does the same.
+    //
+    //
+    // But that's not all! SVG 2 and CSS Inline Layout 3 did a baseline handling overhaul,
+    // and it's far more complex now. Not sure if anyone actually supports it.
+    fn alignment_baseline_shift(&self, alignment: AlignmentBaseline, font_size: f32) -> f32 {
+        match alignment {
+            AlignmentBaseline::Auto => 0.0,
+            AlignmentBaseline::Baseline => 0.0,
+            AlignmentBaseline::BeforeEdge | AlignmentBaseline::TextBeforeEdge => {
+                self.ascent(font_size)
+            }
+            AlignmentBaseline::Middle => self.x_height(font_size) * 0.5,
+            AlignmentBaseline::Central => self.ascent(font_size) - self.height(font_size) * 0.5,
+            AlignmentBaseline::AfterEdge | AlignmentBaseline::TextAfterEdge => {
+                self.descent(font_size)
+            }
+            AlignmentBaseline::Ideographic => self.descent(font_size),
+            AlignmentBaseline::Alphabetic => 0.0,
+            AlignmentBaseline::Hanging => self.ascent(font_size) * 0.8,
+            AlignmentBaseline::Mathematical => self.ascent(font_size) * 0.5,
+        }
+    }
+}
+
+pub(crate) type FontsCache = HashMap<Font, Arc<ResolvedFont>>;
+
+/// A read-only text index in bytes.
+///
+/// Guarantee to be on a char boundary and in text bounds.
+#[derive(Clone, Copy, PartialEq, Debug)]
+pub(crate) struct ByteIndex(usize);
+
+impl ByteIndex {
+    fn new(i: usize) -> Self {
+        ByteIndex(i)
+    }
+
+    pub(crate) fn value(&self) -> usize {
+        self.0
+    }
+
+    /// Converts byte position into a code point position.
+    pub(crate) fn code_point_at(&self, text: &str) -> usize {
+        text.char_indices()
+            .take_while(|(i, _)| *i != self.0)
+            .count()
+    }
+
+    /// Converts byte position into a character.
+    pub(crate) fn char_from(&self, text: &str) -> char {
+        text[self.0..].chars().next().unwrap()
     }
 }
